+++ NEWS.md
@@ -1,6 +1,7 @@
 # admiral.test 0.5.0
  - Updated AE to add variable AELAT for ophthalmology package (#99)
-
+ - Using {metatools} to add labels to datasets (#87)
+ 
 # admiral.test 0.3.0
  - Updated labels for pp and pc datasets (#77)
  - Implemented admiral CI/CD workflows