+++ NEWS.md
@@ -1,6 +1,7 @@
 # admiral.test 0.5.0
  - Updated AE to add variable AELAT for ophthalmology package (#99)
- - Added SDTM SC dataset for ophthalmology package (#102)
+ - Created SDTM SC dataset for ophthalmology package (#102)
+ - Created SDTM OE dataset for ophthalmology package (#101)
 
 # admiral.test 0.3.0
  - Updated labels for pp and pc datasets (#77)