+++ NEWS.md
@@ -3,6 +3,9 @@
  - Created SDTM SC dataset for ophthalmology package (#102)
  - Created SDTM OE dataset for ophthalmology package (#101)
  - Using {metatools} to add labels to datasets (#87)
+ 
+# admiral.test 0.4.0
+ - Various updates to improve CI/CD workflows
 
 # admiral.test 0.3.0
  - Updated labels for pp and pc datasets (#77)